Garapena/Pilotoak/FormakuntzakAPP/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
bitarray==3.0.0
blinker==1.9.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.7
cytoolz==1.0.1
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
eth_abi==5.1.0
Flask==3.1.0
frozenlist==1.5.0
gunicorn==23.0.0
hexbytes==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector==2.2.9
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.0.0
propcache==0.2.1
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.2
pyunormalize==16.0.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rlp==4.0.1
six==1.16.0
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
web3==7.6.1
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
